week2/back/requirements.txt
# blinker==1.9.0
# click==8.2.1
# colorama==0.4.6
# Flask==3.1.1
# itsdangerous==2.2.0
# Jinja2==3.1.6
# MarkupSafe==3.0.2
# python-dotenv==1.1.1
# Werkzeug==3.1.3



flask==2.3.3
psycopg2-binary==2.9.7
python-dotenv==1.0.0
flask-cors==4.0.0